rename ansatz_generation.py -> kak_general.py
